backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==1.10.14
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.5
python-multipart==0.0.6
